Mask interrupt during exec_cmd()
Depending on where we're running, there might be an ISR for HIRQ events,
or something might preempt us and try and run a CD command. Needless to
say, those are all very bad for us, so block them.

If you can convince me that this is a performance issue and you can be
sure you aren't doing anything like this, please let me know, and I'll
add a #define to let you disable it.